Allow proxy web client to be configured via a twisted client endpoint.
Allows you to specify different trust stores for the back channel web client
and the proxy web client.

Flexibility of the new twisted `tls` client endpoint also allows you to
include a client certificate in the proxy web client.  This could be useful
in a scenario such as with OpenShift where the authenticating proxy can
be required to present a client cert in order to access the oauth endpoints.